requirements.txt
flask>=3.0
firebase-admin>=6.5
google-cloud-firestore>=2.16
